rm identifaible and hashable from SwiftData models

+++ Coin/Network/Model/TransactionModels.swift
@@ -25,8 +25,8 @@ struct CreateTransactionRes: Decodable {
 struct UpdateTransactionReq: Encodable {
     var accountFromID: UInt32?
     var accountToID: UInt32?
-    var amountFrom: Double?
-    var amountTo: Double?
+    var amountFrom: Decimal?
+    var amountTo: Decimal?
     var dateTransaction: Date?
     var note: String?
     var id: UInt32
